23/08/13
Long time, no commit. A whole new design, this is the first and last
design (till the beta). Added pjax, ajaxlinks. Now we can use gravatar
OR our avatar storage. Updated the facebook register and login to
prepopulate the fields.

+++ protected/models/Follow.php
@@ -57,6 +57,8 @@ class Follow extends CActiveRecord
             if(!User::model()->findByPk($id) === null)
                 throw new CHttpException(404, 'User to follow not found.');
             $model = $this->findByAttributes(array('user_1'=>Yii::app()->user->id, 'user_2'=>$id));
+            if(empty($model))
+                throw new CHttpException(404, 'You dont follow this user.');
             if($model->delete())
                 return true;
             else
